revert the last few revs, fix a few warnings, add c_str in several locations (still some missing?), fix partial lists
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@662 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ ShareScannerManager.cpp
@@ -324,7 +324,7 @@ bool ShareScannerManager::scanDir(const string& path, int& missingFiles, int& mi
 		}
 	}
 
-	if (!fileList.empty() && ((nfoFiles + sfvFiles) == fileList.size()) && (SETTING(CHECK_EMPTY_RELEASES))) {
+	if (!fileList.empty() && ((nfoFiles + sfvFiles) == (int)fileList.size()) && (SETTING(CHECK_EMPTY_RELEASES))) {
 		reg.assign("(\\S*(((nfo|dir).?fix)|nfo.only)\\S*)", boost::regex_constants::icase);
 		if (!regex_match(dirName,reg)) {
 			folderList = findFiles(path, "*", true);
